+++ docs/history.rst
@@ -9,6 +9,8 @@ Version History
 - Split single module into *logext* and *access* modules
 - The JSON formatter now includes all properties of a ``LogRecord`` instead of 
   hard-coding a fixed set of them.
+- Rename ``JSONRequestFormatter`` to ``JSONFormatter``.  The old name is
+  deprecated.
 
 `1.3.2`_ Oct  2, 2015
 ---------------------